pyproject.toml
[tool.poetry]
name = "sellers-api"
version = "0.1.0"
description = ""
authors = ["Ewald Werner Junior <juninhowerner31@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
djangorestframework = "^3.14.0"
uuid = "^1.30"
validate-docbr = "^1.10.0"
django-filter = "^23.5"
prettyconf = "^2.2.1"
django-cors-headers = "^4.3.1"
gunicorn = "^21.2.0"
psycopg2 = "^2.9.9"
easy-choices = "^1.0.0"
asgiref = "^3.7.2"
boto3 = "^1.34.56"
bcrypt = "^4.1.2"
cryptocode = "^0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
